move keybinding handling to rm-main
