+++ rm-shared/Cargo.toml
@@ -17,4 +17,3 @@ magnetease.workspace = true
 ratatui.workspace = true
 chrono.workspace = true
 serde.workspace = true
-tokio.workspace = true
